Rename all the apps. Ugh.
I realised that not giving the apps `label`s that were unique (eg,
'core' is not likely to be unique) was a big problem. So gave them all
labels, which just breaks everything, including backwards compatibility
with the migrations.

So everything's been renamed, the old new migrations have been replaced.

Bum.