Change URL structure for events
Previously there was a bit of ambiguity between some events and the works
that were featured in them. This was supposed to make things simpler,
and it did in some ways, but also made other things a little more
complicated. And also meant for things like Plays and Movies they had
no individual Event page.

i.e., you couldn't link to a page that
represented when you saw 'Movie X' at a Venue on a date - all those
Events were only listed on the Movie's Detail page. Which is fine,
except when we want to start attaching things (like notes now, maybe
more (pictures?) in the future) to Events.

So we've re-jigged some things, particularly URLs. So we have:

* `/events/` - The list of all recent events.
* `/events/types/movies/` - The list of recent movie Events.
* `/events/<slug>/` - A specific Event of any kind (movie, gig, etc).
* `/events/movies/` - A list of all Movies that have been seen.
* `/events/movies/<slug>/` - A specific Movie.

Similarly for plays.

Classical Works and Dance Pieces (the works themselves, not their
Events) still have URLs like:

* `/events/classical/works/` - A list of all Classical Works.
* `/events/classical/works/<slug>/` - A specific Classical Work.

Simpler kinds of thing, like Gigs, Comedy, Misc do not have separate
URLs because the Event *is* the Gig. There's no "work" attached to
those Events.

For #13