Allow Classical Works and Dance Pieces in an Event to be ordered
Required adding a through model to specify the order of a work
in an Event.

So several migrations to handle that process (create the new m2m
fields with through model; copy the existing data to it; remove the
previous m2m fields; rename the new m2m fields to match the originals).

Added Inline admin fields for these.

Still need to add tests, update Admin JS, change public templates.