Lots of charts scattered around the views/templates
Did start with a separate charts page, but that seemed redundant
once I started adding charts to sidebars etc.

Also added a chartify() method in core.utils for processing a
QuerySet of things in a chart, and giving them a position, based
on their score, allowing for adjacent items with identical scores
which should have the same position.