Don't show months if readings have year granularity
On the year archive pages (i.e. all publications finished in
a particular year) we divide the publications up by the month in which
they were finished.

BUT if we didn't have those detailed dates, and we
only had the year in which a publication was finished (not month or day)
then everything would get listed under "January" for that year.

So now we check, and if the first and last publications read in a year
only have a date granularity of "year", we don't show the months.

Added a "show_months" context variable for the
`reading_archive_year.html` template, which defaults to `True`.

Fixes #41