Don't call geocode APIs on loading the Venue Admin
When loading the Venue add/change page, if there was a lat/lon set,
we were calling the Mapbox/Google geocoding APIs, to fetch the lat/lon
based on the position of the just-positioned marker. Which is a bit
daft.

So now we don't call the geocode APIs in this situation, only doing so
when the marker is repositioned by click or drag.