Fix error with map marker when adding a new venue
Unlike Google Maps, Mabox will add a marker to the map even if we
haven't set its location. Which we only do for a new venue when the user
clicks on the map. So the marker was being added, but with no location
set, and it generated an error that stopped the rest of the JS running.

So now we only add it to the map the first time it has a location
(either because it's an existing venue or when the user first clicks the
map on a new venu).

Fixes #47