Fix position of Mapbox marker when zooming.
The marker would be positioned correctly when the map is very zoomed in,
but as you zoomed it out moved further and further away from the chosen
point. Turns out there's a newish `anchor` option to set when creating
a marker that helps us use the middle-bottom of the marker, to match our
custom icon.

https://stackoverflow.com/a/56323594/250962