Add SPECTATOR_AFFILIATE_CODES setting
There was an undocumented `SPECTATOR_AMAZON` setting available, which
would add Amazon Associates codes to the links to Publications shown
in the `publication_detail` page.

I've now replaced that with a `SPECTATOR_AFFILIATE_CODES` setting
which is more expandable (so we can add bookshop.org codes) and
documented it.

For #56