Fix 'Most read authors' lists
The 'Most read authors' card (and any other lists using that Manager)
was omitting some authors. Not, as I originally thought, if they had
a role of "Author", but if they had *any* Publications that currently
had unfinished Readings.

So, on my site, Samuel Pepys wasn't in the list because I was currently
reading one of his volumes.

Fixes #65