Fix removal of EXIF GPS data when replacing a thumbnail
I thought it worth checking that GPS data was removed from a thumbnail's
EXIF data when we replace an existing thumbnail, not just add one to
a brand new Publication/Event.

Turns out that it wasn't working - the way we were setting
`self.__original_thumbnail` and comparing it to `self.thumbnail`
wasn't working - when we replaced with the new image, both those
values were set to the new image.

I couldn't work out how `__original_thumbnail` was getting set with
the new value just by doing `obj.thumbnail.save(...)`. But it was.

So, for now, we've removed that test and we look for and remove GPS
data from the thumbnail whenever we save the object. Which is...
OKish, but (a) unnecessarily inefficient, and (b) it means that
re-saving an existing `Publication` or `Event` will remove any
current GPS data from its thumbnail. Which *could* be a nice feature
but it's not what you'd expect to happen.

So, should fix that before release.

For #64