Change how a thumbnail is moved
The method I was using to move the location of a thumbnail image
file (so that its path includes the object's `slug`) didn't work
if django-storages was using something like S3, because you can't
rename/move files on that. We got:

    NotImplementedError: This backend doesn't support absolute paths.

So altered the way we handle the file, still ensuring it's saved to
the correct location. Which is actually simpler now. Hopefully it
will still work with S3.

For #67