Use default_storage to load and save thumbnails when editing EXIF
If the project uses something like S3 through django-storages, then
we can't use standard read()ing from FileField paths doesn't work.

So we need to change how we read and save the data when trying to
remove any EXIF data from it, so it will (hopefully) work when
the images are stored off the local filesystem.

For #64