Revert to tox.ini instead of pyproject.toml for tox config
Can't get the tox config working correctly with GitHub Actions.

It seems to run tests across each Django version in turn, for
eacn environment. And if the tests fail for one of the Django
versions, it marks the entire environment as having failed.

e.g,

py311-django42 will run tess using Python 3.11 and Django 4.2,
Django 5.1 and Django main. The latter currently fails, so the
environment py311-django42 will appear as failed, even though
the tests pass with Python 3.11 and Django 4.2.

I created a minimal test repo to try and solve this but had to
give up https://github.com/philgyford/github-actions-tryout

For #89